Split up ZIP modelling folder
