+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/CompetitionZippingData.kt
@@ -1,0 +1,6 @@
+package org.worldcubeassociation.tnoodle.server.webscrambles.zip
+
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.ScrambleDrawingData
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Competition
+
+data class CompetitionZippingData(val wcif: Competition, val namedSheets: Map<String, ScrambleDrawingData>)
